+++ tests/test_activities.py
@@ -1,14 +1,14 @@
 import os
 import warnings
 from datetime import datetime
-from unittest import skipIf
+from unittest import skipIf, skip
 
 import pytest
 import pytz
 import requests
 
 from pykechain.enums import ActivityType, ActivityStatus, ActivityClassification, Category, \
-    activity_root_name_by_classification, ActivityRootNames
+    activity_root_name_by_classification, ActivityRootNames, PaperSize, PaperOrientation, NotificationEvent
 from pykechain.exceptions import NotFoundError, MultipleFoundError, IllegalArgumentError, APIError
 from pykechain.models import Activity2
 from pykechain.utils import temp_chdir
@@ -641,10 +641,7 @@ class TestActivity2SpecificTests(TestBetamax):
                 self.assertTrue(model.property(name='Sale?').output)
         self.assertTrue(len(associated_models) == 3)
 
-
-# @skip('Does not work in PIM2 until KEC-19193 is resolved')
 class TestActivityDownloadAsPDF(TestBetamax):
-
     def test_activity2_download_as_pdf(self):
         # setUp
         activity_name = 'Task - Form'
@@ -662,7 +659,7 @@ class TestActivityDownloadAsPDF(TestBetamax):
     @pytest.mark.skipif("os.getenv('TRAVIS', False) or os.getenv('GITHUB_ACTIONS', False)",
                         reason="Skipping tests when using Travis or Github Actions, as not Auth can be provided")
     def test_activity2_download_as_pdf_async(self):
-        activity_name = 'Task - Form + Tables + Service'
+        activity_name = 'Task - Form'
         activity = self.project.activity(name=activity_name)
 
         # testing
@@ -672,3 +669,57 @@ class TestActivityDownloadAsPDF(TestBetamax):
             pdf_file_called_after_activity = os.path.join(target_dir, activity_name + '.pdf')
             self.assertTrue(pdf_file)
             self.assertTrue(pdf_file_called_after_activity)
+
+    def test_activity2_share_link(self):
+        # setUp
+        test_user = self.client.user(username='testuser')
+
+        activity_name = 'Task - Form'
+        message = 'EXAMPLE_MESSAGE'
+        subject = 'EXAMPLE_SUBJECT'
+        recipient_users = [test_user]
+
+        activity = self.project.activity(name=activity_name)
+
+        activity.share_link(subject=subject,
+                            message=message,
+                            recipient_users=recipient_users)
+
+        # testing
+        notifications = self.client.notifications(subject=subject, message=message,
+                                                  event=NotificationEvent.SHARE_ACTIVITY_LINK)
+        self.assertEqual(self.client.last_response.status_code, requests.codes.ok)
+        self.assertTrue(len(notifications), 1)
+
+        # tearDown
+        notifications[0].delete()
+
+    def test_activity2_share_pdf(self):
+        # setUp
+        test_user = self.client.user(username='testuser')
+
+        activity_name = 'Task - Form'
+        message = 'EXAMPLE_MESSAGE'
+        subject = 'EXAMPLE_SUBJECT'
+        paper_size = PaperSize.A2
+        paper_orientation = PaperOrientation.PORTRAIT
+        recipient_users = [test_user]
+
+        activity = self.project.activity(name=activity_name)
+
+        activity.share_pdf(subject=subject,
+                           message=message,
+                           recipient_users=recipient_users,
+                           paper_size=paper_size,
+                           paper_orientation=paper_orientation,
+                           include_appendices=False
+                           )
+
+        # testing
+        notifications = self.client.notifications(subject=subject, message=message,
+                                                  event=NotificationEvent.SHARE_ACTIVITY_PDF)
+        self.assertEqual(self.client.last_response.status_code, requests.codes.ok)
+        self.assertTrue(len(notifications), 1)
+
+        # tearDown
+        notifications[0].delete()